Adding Functionality to My Connections Page
This starts adding the ability to edit saved connections.  A gear icon has been added to each rendered connection.  When the gear is clicked on then events were added to fire other tags.  In this scenario, the Quick Connect page is loaded and the current connection data is passed through the event.  This will allow for the saved connection data to be loaded into the Quick Connect page and edited.