refs #3 - Delete saved connections
Adding frontend ui elements needed to delete saved connections.  The initial implementation adds flexbox rendering to pages.  This added the ability to align a footer to the bottom of the page.  Additionally, an actionbar was added to the footer implementation which allows for notification popups to be displayed at the bottom of the page.