refs #3, refs #5, sass organization
This pushes connection deletion styles higher in the dom, and starts adding
the ability to delete templates.  The concept of a page variable has been
added to templates and connections.  Additional sass files have been added
for managing variables and functions.  Also, base styles have been broken
into multiple files.  Lastly, the width of the application was changed so
that grids of templates and connections don't leave too much blank spacing
when wrapping.