fixes #7 Build porcelain from gulp
This adds a build and deploy tasks to gulp for managing porcelain builds.
It also adds a watcher for modified python files which will kickoff a
build and deployment.  The previously committed binaries have been removed
and all files within the app/bin directory should be ignored moving
forward.  Lastly, the version number was removed from the porcelain
deployment package.  That application should version with clay and not
independently.