fixes #6 Set text colors
This adds default text colors and a blue override to porcelain.  It also
adds the ability to launch saved connections and read colors from
associated templates.  The bin gitignore file has been replaced by a keep file
as a placeholder for the versioned directory.  Instead of ignoring the
entire bin directory, just the porcelain build directory has been added to
be ignored.